backend/requirements.prod.txt
fastapi[standard]
PyMuPDF
psycopg2